分割したテストファイルを 1ファイルに戻す
